requirements.txt
RPi.GPIO==0.7.0
pillow==7.0.0
gfxcili==0.2
gfxlcd==0.8.3
spidev==3.4
